did some testing, fixed a couple matrix errors

+++ newfiles/randos.h
@@ -28,7 +28,7 @@ static int random_int(int max)
 }
 
 // OUTPUT: a pseudo-random float between the range 0 and 999999999
-static int random_float() {
+static float random_float() {
     static std::mt19937 m_mt( std::time(nullptr) );
     static std::uniform_real_distribution<float> dist(0, 999999999);
 
@@ -37,7 +37,7 @@ static int random_float() {
 
 // INPUT: a maximum bound to the range used to generate a pseudorandom float
 // OUTPUT: a pseudorandom float between 0 and max
-static int random_float(int max) {
+static float random_float(int max) {
     static std::mt19937 m_mt( std::time(nullptr) );
     static std::uniform_real_distribution<float> dist(0, max);
 
@@ -45,7 +45,7 @@ static int random_float(int max) {
 }
 
 // OUTPUT: a pseudo-random double between the range 0 and 999999999
-static int random_double()
+static double random_double()
 {
     static std::mt19937 m_mt( std::time(nullptr) );
     static std::uniform_real_distribution<double> dist(0, 999999999);
@@ -55,7 +55,7 @@ static int random_double()
 
 // INPUT: a maximum bound to the range used to generate a pseudorandom double
 // OUTPUT: a pseudorandom double between 0 and max
-static int random_double(int max)
+static double random_double(int max)
 {
     static std::mt19937 m_mt( std::time(nullptr) );
     static std::uniform_real_distribution<double> dist(0, max);
